requirements.txt
Python==3.10.8
folium
GeoAlchemy2
geojson
geopandas
pandas
Pillow
plotly
psycopg2
pytest
pytest-xdist
python-dotenv
SQLAlchemy
SQLAlchemy-Utils
streamlit
streamlit-folium
vcrpy

